requirements.txt
streamlit==1.32.2
Pillow==10.2.0
vertexai==0.3.3
google-cloud-aiplatform==1.40.0
markdown2==2.4.10
matplotlib==3.8.3
numpy==1.26.4
python-dotenv==1.0.1
google-auth==2.28.1
requests==2.31.0
